Created another DataDomain project to remove circular dependency between projects MathsVisualisationTool and Parser
