+++ MathsVisualisationTool.sln
@@ -4,8 +4,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 15.0.28307.421
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathsVisualisationTool", "MathsVisualisationTool\MathsVisualisationTool.csproj", "{0146A510-5F39-4F11-AFBC-C55338B15B63}"
+	ProjectSection(ProjectDependencies) = postProject
+		{354F4FA4-92B0-4285-8B4B-EA3A8656C701} = {354F4FA4-92B0-4285-8B4B-EA3A8656C701}
+	EndProjectSection
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Parser", "Parser\Parser.fsproj", "{354F4FA4-92B0-4285-8B4B-EA3A8656C701}"
+	ProjectSection(ProjectDependencies) = postProject
+		{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF} = {E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataDomain", "DataDomain\DataDomain.csproj", "{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +29,10 @@ Global
 		{354F4FA4-92B0-4285-8B4B-EA3A8656C701}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{354F4FA4-92B0-4285-8B4B-EA3A8656C701}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{354F4FA4-92B0-4285-8B4B-EA3A8656C701}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E99A2845-A6AB-48AA-ACFD-16FC9AFCE5CF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
