+++ MathsVisualisationTool/interpreter/Interpreter.cs
@@ -3,7 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Parser.FSharpImplementation2;
+using Parser.FSharpImplementation;
+using DataDomain;
 
 namespace MathsVisualisationTool
 {
@@ -26,8 +27,6 @@ namespace MathsVisualisationTool
             //Check if the syntax is okay - throws an error if not and 
             //runs the code.
             double result = parser.AnalyseTokens(tokens);
-
-            Example.add(1, 2);
 
             return Convert.ToString(result);
         }
