+++ Parser/Parser.fs
@@ -1,0 +1,13 @@
+﻿// Learn more about F# at http://fsharp.org
+// See the 'F# Tutorial' project for more help.
+
+namespace Parser.FSharpImplementation
+open DataDomain;
+open System.Collections;
+
+module Expression =
+    (*let expression tokens:List<Token> =
+        let term =
+            
+        let expression1 = *) 
+
