+++ Parser/Parser.fsproj
@@ -48,7 +48,7 @@
   <Import Project="$(FSharpTargetsPath)" />
   <ItemGroup>
     <Compile Include="AssemblyInfo.fs" />
-    <Compile Include="Program.fs" />
+    <Compile Include="Parser.fs" />
     <None Include="App.config" />
     <Content Include="packages.config" />
   </ItemGroup>
