--- Parser/Program.fs
@@ -1,8 +1,0 @@
-﻿// Learn more about F# at http://fsharp.org
-// See the 'F# Tutorial' project for more help.
-
-namespace Parser.FSharpImplementation2
-
-module Example =
-    let add num1 num2 = num1 + num2
-
